--- packages/zod-request/src/ZodErrorSerializer.ts
@@ -1,4 +1,0 @@
-import type z from 'zod';
-export const serializeZodError = (_error: z.ZodError) => {
-  return 'cool';
-};
